requirements.txt
requests
pandas
pandas_ta
numpy
catboost
flask
gunicorn
matplotlib
scikit-learn
python-dotenv
